now it is possible to choose the vairant when deleting project feature

+++ src/Controller/ProjectController.php
@@ -116,7 +116,7 @@ class ProjectController extends AbstractController
     }
 
     /**
-     * @Route("/{id}/{variant<high|middle|low>}", name="project_delete", methods={"DELETE"})
+     * @Route("/{id}", name="project_delete", methods={"DELETE"})
      */
     public function delete(Request $request, Project $project, string $variant = 'high'): Response
     {
@@ -130,23 +130,27 @@ class ProjectController extends AbstractController
     }
 
     /**
-     * @Route("Feature/{id}", name="project_feature_delete", methods="POST")
+     * @Route("Feature/{id}/{variant<high|middle|low>}", name="project_feature_delete", methods="POST")
      * @param ProjectFeature         $projectFeature
      * @param EntityManagerInterface $entityManager
      * @return Response
      */
     public function deleteProjectFeature(
         ProjectFeature $projectFeature,
-        EntityManagerInterface $entityManager
+        EntityManagerInterface $entityManager,
+        ProjectCalculator $projectCalculator,
+        string $variant = 'high'
     ): Response {
-        $entityManager->remove($projectFeature);
+        $variant=ucfirst($variant);
+        $projectFeature->{'setIs'.$variant}(false);
+        if (!$projectCalculator->isAlive($projectFeature)) {
+            $entityManager->remove($projectFeature);
+        }
         $entityManager->flush();
-
         /** @var Project */
         $project = $projectFeature->getProject();
         $projectId = $project->getId();
-
-        return $this->redirectToRoute('project_edit', ['id' => $projectId]);
+        return $this->redirectToRoute('project_edit', ['id' => $projectId, 'estimation'=>$variant]);
     }
 
     /**
